CalAI-Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-ai-slim[openai,google]==1.0.18
google-generativeai==0.8.3
supabase==2.22.0
logfire==2.6.0
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20
pillow==11.3.0